+++ repos/sbt/sbt/src/sbt-test/tests/done/src/main/scala/custom/CustomReporter.scala
@@ -9,10 +9,8 @@ class CustomReporter extends ResourcefulReporter {
   private def writeFile(filePath: String, content: String): Unit = {
     val file = new File(filePath)
     val writer =
-      if (!file.exists)
-        new FileWriter(new File(filePath))
-      else
-        new FileWriter(new File(filePath + "-2"))
+      if (!file.exists) new FileWriter(new File(filePath))
+      else new FileWriter(new File(filePath + "-2"))
     writer.write(content)
     writer.flush()
     writer.close()
@@ -29,10 +27,7 @@ class CustomReporter extends ResourcefulReporter {
   def dispose(): Unit = {
     val file = new File("target/dispose")
     val filePath =
-      if (file.exists)
-        "target/dispose2"
-      else
-        "target/dispose"
+      if (file.exists) "target/dispose2" else "target/dispose"
     writeFile(filePath, "dispose")
   }
 }
